refactor(conductor): extract applyIntent and card factories from turn-handler
Move 7 applyIntent* functions to card-mutations.ts.
Move createEmpty* factories + modeToCardType to card-factories.ts.
turn-handler.ts reduced from ~780 to ~140 lines.
Re-exports maintain backward compatibility for all existing imports.
Closes #173
